Move Traversing utils to detekt-rules-documentation
