+++ detekt-rules-documentation/src/main/kotlin/io/gitlab/arturbosch/detekt/rules/documentation/internal/Traversing.kt
@@ -1,5 +1,6 @@
-package io.gitlab.arturbosch.detekt.rules
+package io.gitlab.arturbosch.detekt.rules.documentation.internal
 
+import io.gitlab.arturbosch.detekt.rules.isProtected
 import org.jetbrains.kotlin.psi.KtNamedDeclaration
 import org.jetbrains.kotlin.psi.psiUtil.containingClassOrObject
 import org.jetbrains.kotlin.psi.psiUtil.isPublic
